Improve handling of unitig mappings
A UnitigMapping object can represent the mapping of a single k-mer
on a unitig, of a mapping of the full unitig in itself.

The unitig iterator on a graph always returns full mappings, when
iterating over the successors or predecessors you'll always get
full mappings. And now, with this commit, when retreiving a unitig
through `g.nodes`, you'll also get a full mapping.

See https://github.com/pmelsted/bifrost/issues/17