Make k-mer more central to the API, huge refactor
This commit includes a big refactor to the API to make the K-mer
more central. In the previous version, nodes were represented by a
`PyfrostColoredUMap` objects. Now, each node is identified by its
"head" k-mer.

This has several benefits: Kmer objects are more light weight, and
are immutable objects. Previously, the PyfrostColoredUMap objects
all contained references to the original graph, which is recipe for
segfaults and memory leaks, especially when in a later stage we add
the possibility to modify the De Bruijn graph.

Furthermore, the API to access nodes and search for k-mers feels
more Pythonic and natural.

This commit also includes a `DegreeView` class, so obtaining node
degrees is now possible too.