Refactor link annotation for a specific color
Link annotation for a specific color now has its own class. This
makes the code more generic, and fixes a bug in the old code that
could match a k-mer on a unitig which was not associated with the
color of interest.

The associated color will be stored along side all junction trees
when saving a MemLinkDB to a file.