Add 'max_link_length' option to LinkAnnotator
This is useful for reference genomes: when adding links from
long contigs many nodes in the graph will be annotated with
long links (because the path of the contig is very long, and thus
traverses many junctions).

In many cases such long links are not necessary, because you'll
only need links long enough to span the largest repeat, because
the algorithm will pick up new links along the way.