Improve link annotation algorithm
Previously, annotation would stop if k-mers didn't match or if
there was any ambiguity. Now, instead of quiting, we reset which
nodes to annotate and allow annotation to continue later.

This is especially useful for long sequences (e.g., reference
genomes), where parts of the genome may not match with the graph,
but other parts do match, now allowing annotation to continue even
if earlier parts mismatched.