Change isLeaf() in JunctionTree
If the coverage of a JunctionTreeNode is higher than the sum of its
children, then it means certain links end at that node, and
technically they are leaf nodes.

We are, however, generally not interested in these shorter links,
and with the changed isLeaf() function these shorter links are now
not considered leafs anymore. This results in fewer links picked
up at a node, and thus shorter runtime when navigating.